+++ NEWS.md
@@ -15,7 +15,7 @@
 - RStudio's GWT sources can now be built with JDKs > 11 (#11242)
 - Show grey background instead of solid-white during Desktop startup (#13768)
 - The 'restartSession()' API method gains the 'clean' argument. (#2841)
-- 'dot' and 'mermaid' chunks in R Markdown documents are now executable (#14063)
+- 'dot' chunks in R Markdown documents are now executable (#14063)
 
 #### Posit Workbench
 - Show custom project names on Workbench homepage (rstudio-pro#5589)